Fix uncaught exception
The bug was that if the user enters an expression into an Input or Want
field, and the expression parses correctly but causes an interpreter
error (such as an undefined variable name, or (+ 2 true)), then an
uncaught exception breaks the application.  Whereas for output formulas
we show interpreter errors to the user, for Input and Want we do not
currently show interpreter errors to the user, so we should consider
those errors to make the expression invalid.