implemented colors, implemented solid and outlined circles, rects
Outlined shapes are kind of weird. This is because the "stroke"
attribute of svg shapes can't make up its mind of whether it wants to
draw outside a shape or inside a shape, so it does a bit of both. Also
I'm not sure if this behavior exists on browsers besides firefox.

Also I still haven't tested any of this so I'm still not sure it works