kind of implemented define-struct
there's a function makeStruct that returns an environment with the
functions that we would expect define-struct would make. However,
there's currently no way to actually access this function from the table
itself. I'm not quite sure how to actually implement this so someone can
actually make their own structs that they can use in the table.

Anyways, I tested makeStruct using the point structure as an example and
it seems to work.