(define-struct ...) now works
added prefix field (currently shows up below the language selector box)
where one can write expressions that change the environment that the
tables use to interpret expressions. Currently only (define-struct
...) works here and not (define ...)

Also changed how Input and Want fields are validated. Previously they
needed to contain expressions that could be parsed and interpreted
without throwing an error in able to be considered valid (and thus
change the state of the table). Now they only need to be able to be
parsed without throwing an error for them to be considered valid. Due to
this an Input or a Want can contain an expression that parses, but
throws an error. This error is caught and shown to the user in both
cases. For Wants it is shown to the right (where images are shown if the
expression in the Want evaluates to an image). For Inputs it is shown
below the Input. I'm not sure this is the best place to put Input
errors, it just made sense to me at the time. However the location of
Input errors may be subject to change in the future.

Also the location of the Prefix textarea should probably be changed
because it looks kind of awkward right now.