wrote playback.js
haven't tested whether it actually works though, have to figure out a
mechanism to fetch log files (the files that include the JSON that
represents the list of snapshots of a particular instance of the
beginning student tables) from the log server.

Maybe have a menu somewhere that shows all the log files at the server
and you can pick one of them.